feat: 添加 Conv BatchNormalization Pool 和 GlobalPool 的推导

+++ src/04frontend/include/frontend/operator.h
@@ -8,8 +8,8 @@
 
 namespace refactor::frontend {
 
-    using Int = long long;
-    using Ints = std::vector<long long>;
+    using Int = int64_t;
+    using Ints = std::vector<int64_t>;
     using Float = float;
     using Floats = std::vector<float>;
     using String = std::string;
